Rename --pattern to --include and --exclude; refactor depfilter
