+++ crates/csvizmo-depgraph/src/algorithm/select.rs
@@ -1,18 +1,23 @@
-use std::collections::HashSet;
+use std::collections::{HashSet, VecDeque};
 
 use clap::Parser;
 use petgraph::Direction;
+use petgraph::graph::NodeIndex;
 
 use super::{MatchKey, build_globset};
-use crate::{DepGraph, FlatGraphView};
+use crate::{DepGraph, Edge, FlatGraphView};
 
 #[derive(Clone, Debug, Default, Parser)]
 pub struct SelectArgs {
-    /// Glob pattern to select nodes (can be repeated)
-    #[clap(short, long)]
-    pub pattern: Vec<String>,
+    /// Glob pattern to include nodes (can be repeated)
+    #[clap(short = 'g', long)]
+    pub include: Vec<String>,
 
-    /// Combine multiple patterns with AND instead of OR
+    /// Glob pattern to exclude nodes (can be repeated, always OR)
+    #[clap(short = 'x', long)]
+    pub exclude: Vec<String>,
+
+    /// Combine multiple include patterns with AND instead of OR
     #[clap(long)]
     pub and: bool,
 
@@ -31,11 +36,21 @@ pub struct SelectArgs {
     /// Traverse up to N layers (implies --deps if no direction given)
     #[clap(long)]
     pub depth: Option<usize>,
+
+    /// Preserve graph connectivity when excluding nodes
+    /// (creates direct edges, no self-loops or parallel edges)
+    #[clap(long)]
+    pub preserve_connectivity: bool,
 }
 
 impl SelectArgs {
-    pub fn pattern(mut self, p: impl Into<String>) -> Self {
-        self.pattern.push(p.into());
+    pub fn include(mut self, p: impl Into<String>) -> Self {
+        self.include.push(p.into());
+        self
+    }
+
+    pub fn exclude(mut self, p: impl Into<String>) -> Self {
+        self.exclude.push(p.into());
         self
     }
 
@@ -63,21 +78,34 @@ impl SelectArgs {
         self.depth = Some(n);
         self
     }
+
+    pub fn preserve_connectivity(mut self) -> Self {
+        self.preserve_connectivity = true;
+        self
+    }
 }
 
 pub fn select(graph: &DepGraph, args: &SelectArgs) -> eyre::Result<DepGraph> {
-    let globset = build_globset(&args.pattern)?;
     let view = FlatGraphView::new(graph);
 
     // No filters at all -> pass through the entire graph unchanged.
     let no_traversal = !args.deps && !args.rdeps && args.depth.is_none();
-    if args.pattern.is_empty() && no_traversal {
+    if args.include.is_empty() && args.exclude.is_empty() && no_traversal {
         return Ok(graph.clone());
     }
 
-    // If no patterns given, seed from root nodes; otherwise match by pattern.
-    let mut keep: HashSet<_> = if args.pattern.is_empty() {
-        view.roots().collect()
+    let include_globset = build_globset(&args.include)?;
+    let exclude_globset = build_globset(&args.exclude)?;
+
+    // Build the initial keep set from --include patterns (or all nodes / roots).
+    let has_traversal = args.deps || args.rdeps || args.depth.is_some();
+    let mut keep: HashSet<_> = if args.include.is_empty() {
+        if has_traversal {
+            view.roots().collect()
+        } else {
+            // Only --exclude given, no traversal: start with all nodes.
+            view.id_to_idx.values().copied().collect()
+        }
     } else {
         let mut matched = HashSet::new();
         for (id, info) in graph.all_nodes() {
@@ -87,9 +115,9 @@ pub fn select(graph: &DepGraph, args: &SelectArgs) -> eyre::Result<DepGraph> {
             };
 
             let is_match = if args.and {
-                globset.matches(text).len() == args.pattern.len()
+                include_globset.matches(text).len() == args.include.len()
             } else {
-                globset.is_match(text)
+                include_globset.is_match(text)
             };
 
             if is_match && let Some(&idx) = view.id_to_idx.get(id.as_str()) {
@@ -99,6 +127,7 @@ pub fn select(graph: &DepGraph, args: &SelectArgs) -> eyre::Result<DepGraph> {
         matched
     };
 
+    // Expand keep set via --deps/--rdeps/--depth (only applies to include).
     // --depth without an explicit direction implies --deps
     let deps = args.deps || args.depth.is_some();
     if deps && args.rdeps {
@@ -111,7 +140,112 @@ pub fn select(graph: &DepGraph, args: &SelectArgs) -> eyre::Result<DepGraph> {
         keep = view.bfs(keep, Direction::Outgoing, args.depth);
     }
 
-    Ok(view.filter(&keep))
+    // Remove nodes matching --exclude patterns from keep set.
+    let excluded: HashSet<NodeIndex> = if !args.exclude.is_empty() {
+        let mut matched = HashSet::new();
+        for (id, info) in graph.all_nodes() {
+            let text = match args.key {
+                MatchKey::Id => id.as_str(),
+                MatchKey::Label => info.label.as_str(),
+            };
+
+            if exclude_globset.is_match(text)
+                && let Some(&idx) = view.id_to_idx.get(id.as_str())
+            {
+                matched.insert(idx);
+            }
+        }
+        let excluded = keep.intersection(&matched).copied().collect::<HashSet<_>>();
+        for &idx in &excluded {
+            keep.remove(&idx);
+        }
+        excluded
+    } else {
+        HashSet::new()
+    };
+
+    let mut result = view.filter(&keep);
+
+    // Bypass excluded nodes: connect their surviving predecessors to surviving successors.
+    // BFS through chains of excluded nodes so that A->B->C->D with B,C excluded produces A->D.
+    if args.preserve_connectivity && !excluded.is_empty() {
+        let mut existing: HashSet<(String, String)> = result
+            .all_edges()
+            .iter()
+            .map(|e| (e.from.clone(), e.to.clone()))
+            .collect();
+
+        let mut bypass_edges = Vec::new();
+        for &idx in &excluded {
+            let preds = surviving_neighbors(&view.pg, idx, Direction::Incoming, &keep);
+            let succs = surviving_neighbors(&view.pg, idx, Direction::Outgoing, &keep);
+
+            for &pred in &preds {
+                let from = view.idx_to_id[pred.index()];
+                for &succ in &succs {
+                    let to = view.idx_to_id[succ.index()];
+                    if from != to && existing.insert((from.to_string(), to.to_string())) {
+                        bypass_edges.push((from.to_string(), to.to_string()));
+                    }
+                }
+            }
+        }
+
+        for (from, to) in bypass_edges {
+            insert_edge(&mut result, &from, &to);
+        }
+
+        result.clear_caches();
+    }
+
+    Ok(result)
+}
+
+/// Insert a bypass edge into the deepest subgraph that contains both endpoints.
+/// Falls back to the root graph if the endpoints are in different subgraphs.
+fn insert_edge(graph: &mut DepGraph, from: &str, to: &str) {
+    for sg in &mut graph.subgraphs {
+        let has_from = sg.all_nodes().contains_key(from);
+        let has_to = sg.all_nodes().contains_key(to);
+        if has_from && has_to {
+            return insert_edge(sg, from, to);
+        }
+    }
+    graph.edges.push(Edge {
+        from: from.to_string(),
+        to: to.to_string(),
+        ..Default::default()
+    });
+}
+
+/// BFS from `start` in `direction`, traversing through removed nodes (those not in `keep`),
+/// returning surviving nodes found at the boundary.
+fn surviving_neighbors(
+    pg: &petgraph::Graph<(), ()>,
+    start: NodeIndex,
+    direction: Direction,
+    keep: &HashSet<NodeIndex>,
+) -> Vec<NodeIndex> {
+    let mut result = Vec::new();
+    let mut visited = HashSet::new();
+    let mut queue = VecDeque::new();
+    visited.insert(start);
+    queue.push_back(start);
+
+    while let Some(node) = queue.pop_front() {
+        for neighbor in pg.neighbors_directed(node, direction) {
+            if !visited.insert(neighbor) {
+                continue;
+            }
+            if keep.contains(&neighbor) {
+                result.push(neighbor);
+            } else {
+                queue.push_back(neighbor);
+            }
+        }
+    }
+
+    result
 }
 
 #[cfg(test)]
@@ -154,7 +288,7 @@ mod tests {
             .collect()
     }
 
-    // -- pattern matching --
+    // -- include pattern matching --
 
     #[test]
     fn single_glob_pattern() {
@@ -172,7 +306,7 @@ mod tests {
             ],
             vec![],
         );
-        let args = SelectArgs::default().pattern("lib*");
+        let args = SelectArgs::default().include("lib*");
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["libfoo", "libbar"]);
         assert_eq!(edge_pairs(&result), vec![("libfoo", "libbar")]);
@@ -181,7 +315,7 @@ mod tests {
     #[test]
     fn match_by_id() {
         let g = make_graph(&[("1", "libfoo"), ("2", "libbar")], &[("1", "2")], vec![]);
-        let args = SelectArgs::default().pattern("1").key(MatchKey::Id);
+        let args = SelectArgs::default().include("1").key(MatchKey::Id);
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["1"]);
     }
@@ -189,7 +323,7 @@ mod tests {
     #[test]
     fn match_by_label() {
         let g = make_graph(&[("1", "libfoo"), ("2", "libbar")], &[("1", "2")], vec![]);
-        let args = SelectArgs::default().pattern("libbar");
+        let args = SelectArgs::default().include("libbar");
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["2"]);
     }
@@ -201,7 +335,7 @@ mod tests {
             &[("a", "b"), ("b", "c")],
             vec![],
         );
-        let args = SelectArgs::default().pattern("a").pattern("c");
+        let args = SelectArgs::default().include("a").include("c");
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["a", "c"]);
         assert!(edge_pairs(&result).is_empty());
@@ -219,8 +353,8 @@ mod tests {
             vec![],
         );
         let args = SelectArgs::default()
-            .pattern("libfoo*")
-            .pattern("*alpha")
+            .include("libfoo*")
+            .include("*alpha")
             .and();
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["libfoo-alpha"]);
@@ -229,7 +363,7 @@ mod tests {
     #[test]
     fn no_match_produces_empty_graph() {
         let g = make_graph(&[("a", "a"), ("b", "b")], &[("a", "b")], vec![]);
-        let args = SelectArgs::default().pattern("nonexistent");
+        let args = SelectArgs::default().include("nonexistent");
         let result = select(&g, &args).unwrap();
         assert!(result.nodes.is_empty());
         assert!(result.edges.is_empty());
@@ -245,7 +379,7 @@ mod tests {
             &[("a", "b"), ("b", "c"), ("a", "c")],
             vec![],
         );
-        let args = SelectArgs::default().pattern("a").deps();
+        let args = SelectArgs::default().include("a").deps();
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["a", "b", "c"]);
     }
@@ -258,7 +392,7 @@ mod tests {
             &[("a", "b"), ("b", "c")],
             vec![],
         );
-        let args = SelectArgs::default().pattern("c").rdeps();
+        let args = SelectArgs::default().include("c").rdeps();
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["a", "b", "c"]);
     }
@@ -271,7 +405,7 @@ mod tests {
             &[("a", "b"), ("b", "c"), ("c", "d")],
             vec![],
         );
-        let args = SelectArgs::default().pattern("a").deps().depth(1);
+        let args = SelectArgs::default().include("a").deps().depth(1);
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["a", "b"]);
         assert_eq!(edge_pairs(&result), vec![("a", "b")]);
@@ -285,7 +419,7 @@ mod tests {
             &[("a", "b"), ("b", "c"), ("c", "d")],
             vec![],
         );
-        let args = SelectArgs::default().pattern("b").deps().rdeps();
+        let args = SelectArgs::default().include("b").deps().rdeps();
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["a", "b", "c", "d"]);
         assert_eq!(
@@ -302,7 +436,7 @@ mod tests {
             &[("a", "b"), ("b", "c"), ("c", "d"), ("d", "e")],
             vec![],
         );
-        let args = SelectArgs::default().pattern("c").deps().rdeps().depth(1);
+        let args = SelectArgs::default().include("c").deps().rdeps().depth(1);
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["b", "c", "d"]);
         assert_eq!(edge_pairs(&result), vec![("b", "c"), ("c", "d")]);
@@ -362,10 +496,248 @@ mod tests {
             &[("a", "b"), ("b", "c")],
             vec![make_graph(&[("c", "c")], &[], vec![])],
         );
-        let args = SelectArgs::default().pattern("a").deps();
+        let args = SelectArgs::default().include("a").deps();
         let result = select(&g, &args).unwrap();
         assert_eq!(node_ids(&result), vec!["a", "b"]);
         assert_eq!(result.subgraphs.len(), 1);
         assert_eq!(node_ids(&result.subgraphs[0]), vec!["c"]);
+    }
+
+    // -- exclude pattern matching --
+
+    #[test]
+    fn exclude_single_pattern() {
+        // myapp -> libfoo -> libbar, myapp -> libbar
+        let g = make_graph(
+            &[
+                ("libfoo", "libfoo"),
+                ("libbar", "libbar"),
+                ("myapp", "myapp"),
+            ],
+            &[
+                ("myapp", "libfoo"),
+                ("myapp", "libbar"),
+                ("libfoo", "libbar"),
+            ],
+            vec![],
+        );
+        let args = SelectArgs::default().exclude("libfoo");
+        let result = select(&g, &args).unwrap();
+        assert_eq!(node_ids(&result), vec!["libbar", "myapp"]);
+        assert_eq!(edge_pairs(&result), vec![("myapp", "libbar")]);
+    }
+
+    #[test]
+    fn exclude_multiple_patterns_or() {
+        let g = make_graph(
+            &[("a", "a"), ("b", "b"), ("c", "c")],
+            &[("a", "b"), ("b", "c")],
+            vec![],
+        );
+        let args = SelectArgs::default().exclude("a").exclude("b");
+        let result = select(&g, &args).unwrap();
+        assert_eq!(node_ids(&result), vec!["c"]);
+        assert!(edge_pairs(&result).is_empty());
+    }
+
+    #[test]
+    fn exclude_no_match_returns_unchanged() {
+        let g = make_graph(&[("a", "a"), ("b", "b")], &[("a", "b")], vec![]);
+        let args = SelectArgs::default().exclude("nonexistent");
+        let result = select(&g, &args).unwrap();
+        assert_eq!(node_ids(&result), vec!["a", "b"]);
+        assert_eq!(edge_pairs(&result), vec![("a", "b")]);
+    }
+
+    // -- exclude with preserve connectivity --
+
+    #[test]
+    fn preserve_connectivity_bypass() {
+        // a -> b -> c: exclude b, get a -> c
+        let g = make_graph(
+            &[("a", "a"), ("b", "b"), ("c", "c")],
+            &[("a", "b"), ("b", "c")],
+            vec![],
+        );
+        let args = SelectArgs::default().exclude("b").preserve_connectivity();
+        let result = select(&g, &args).unwrap();
+        assert_eq!(node_ids(&result), vec!["a", "c"]);
+        assert_eq!(edge_pairs(&result), vec![("a", "c")]);
+    }
+
+    #[test]
+    fn preserve_connectivity_chain() {
+        // a -> b -> c -> d: exclude b and c, get a -> d
+        let g = make_graph(
+            &[("a", "a"), ("b", "b"), ("c", "c"), ("d", "d")],
+            &[("a", "b"), ("b", "c"), ("c", "d")],
+            vec![],
+        );
+        let args = SelectArgs::default()
+            .exclude("b")
+            .exclude("c")
+            .preserve_connectivity();
+        let result = select(&g, &args).unwrap();
+        assert_eq!(node_ids(&result), vec!["a", "d"]);
+        assert_eq!(edge_pairs(&result), vec![("a", "d")]);
+    }
+
+    #[test]
+    fn preserve_connectivity_diamond_through_excluded() {
+        // a -> b -> d, a -> c -> d: exclude b and c, get single a -> d
+        let g = make_graph(
+            &[("a", "a"), ("b", "b"), ("c", "c"), ("d", "d")],
+            &[("a", "b"), ("a", "c"), ("b", "d"), ("c", "d")],
+            vec![],
+        );
+        let args = SelectArgs::default()
+            .exclude("b")
+            .exclude("c")
+            .preserve_connectivity();
+        let result = select(&g, &args).unwrap();
+        assert_eq!(node_ids(&result), vec!["a", "d"]);
+        assert_eq!(edge_pairs(&result), vec![("a", "d")]);
+    }
+
+    #[test]
+    fn preserve_connectivity_no_self_loops() {
+        // a -> b -> a: exclude b, should not create a -> a
+        let g = make_graph(&[("a", "a"), ("b", "b")], &[("a", "b"), ("b", "a")], vec![]);
+        let args = SelectArgs::default().exclude("b").preserve_connectivity();
+        let result = select(&g, &args).unwrap();
+        assert_eq!(node_ids(&result), vec!["a"]);
+        assert!(edge_pairs(&result).is_empty());
+    }
+
+    #[test]
+    fn preserve_connectivity_no_parallel_edges() {
+        // a -> b -> c, a -> c: exclude b, should not duplicate a -> c
+        let g = make_graph(
+            &[("a", "a"), ("b", "b"), ("c", "c")],
+            &[("a", "b"), ("b", "c"), ("a", "c")],
+            vec![],
+        );
+        let args = SelectArgs::default().exclude("b").preserve_connectivity();
+        let result = select(&g, &args).unwrap();
+        assert_eq!(node_ids(&result), vec!["a", "c"]);
+        assert_eq!(edge_pairs(&result), vec![("a", "c")]);
+    }
+
+    // -- exclude with subgraphs --
+
+    #[test]
+    fn exclude_preserves_subgraph_structure() {
+        // root: a, subgraph: { b, c, b->c }, edge a->b at root
+        // exclude a keeps b, c in subgraph with their edge
+        let g = make_graph(
+            &[("a", "a")],
+            &[("a", "b")],
+            vec![make_graph(&[("b", "b"), ("c", "c")], &[("b", "c")], vec![])],
+        );
+        let args = SelectArgs::default().exclude("a");
+        let result = select(&g, &args).unwrap();
+        assert!(result.nodes.is_empty());
+        assert!(result.edges.is_empty());
+        assert_eq!(result.subgraphs.len(), 1);
+        assert_eq!(node_ids(&result.subgraphs[0]), vec!["b", "c"]);
+        assert_eq!(edge_pairs(&result.subgraphs[0]), vec![("b", "c")]);
+    }
+
+    // -- preserve connectivity with subgraphs --
+
+    #[test]
+    fn preserve_connectivity_bypass_in_subgraph() {
+        // subgraph { a -> b -> c }: exclude b, bypass a -> c should be in the subgraph
+        let g = make_graph(
+            &[],
+            &[],
+            vec![make_graph(
+                &[("a", "a"), ("b", "b"), ("c", "c")],
+                &[("a", "b"), ("b", "c")],
+                vec![],
+            )],
+        );
+        let args = SelectArgs::default().exclude("b").preserve_connectivity();
+        let result = select(&g, &args).unwrap();
+        // bypass a -> c should be inside the subgraph, not at root
+        assert!(result.edges.is_empty());
+        assert_eq!(result.subgraphs.len(), 1);
+        let sg = &result.subgraphs[0];
+        assert_eq!(node_ids(sg), vec!["a", "c"]);
+        assert_eq!(edge_pairs(sg), vec![("a", "c")]);
+    }
+
+    #[test]
+    fn preserve_connectivity_no_parallel_edges_in_subgraph() {
+        // subgraph { a -> b -> c, a -> c }: exclude b, should not duplicate a -> c
+        let g = make_graph(
+            &[],
+            &[],
+            vec![make_graph(
+                &[("a", "a"), ("b", "b"), ("c", "c")],
+                &[("a", "b"), ("b", "c"), ("a", "c")],
+                vec![],
+            )],
+        );
+        let args = SelectArgs::default().exclude("b").preserve_connectivity();
+        let result = select(&g, &args).unwrap();
+        assert!(result.edges.is_empty());
+        assert_eq!(result.subgraphs.len(), 1);
+        let sg = &result.subgraphs[0];
+        assert_eq!(node_ids(sg), vec!["a", "c"]);
+        assert_eq!(edge_pairs(sg), vec![("a", "c")]);
+    }
+
+    #[test]
+    fn preserve_connectivity_cross_subgraph_bypass_at_root() {
+        // subgraph1 { a }, subgraph2 { c }, root: b, edges a->b, b->c at root
+        // exclude b, bypass a->c should be at root
+        let g = make_graph(
+            &[("b", "b")],
+            &[("a", "b"), ("b", "c")],
+            vec![
+                make_graph(&[("a", "a")], &[], vec![]),
+                make_graph(&[("c", "c")], &[], vec![]),
+            ],
+        );
+        let args = SelectArgs::default().exclude("b").preserve_connectivity();
+        let result = select(&g, &args).unwrap();
+        assert!(result.nodes.is_empty());
+        assert_eq!(edge_pairs(&result), vec![("a", "c")]);
+        assert_eq!(result.subgraphs.len(), 2);
+    }
+
+    // -- combined include + exclude --
+
+    #[test]
+    fn include_with_exclude() {
+        // a -> b -> c -> d: include a with deps, then exclude c
+        let g = make_graph(
+            &[("a", "a"), ("b", "b"), ("c", "c"), ("d", "d")],
+            &[("a", "b"), ("b", "c"), ("c", "d")],
+            vec![],
+        );
+        let args = SelectArgs::default().include("a").deps().exclude("c");
+        let result = select(&g, &args).unwrap();
+        assert_eq!(node_ids(&result), vec!["a", "b", "d"]);
+        assert_eq!(edge_pairs(&result), vec![("a", "b")]);
+    }
+
+    #[test]
+    fn include_with_exclude_preserve_connectivity() {
+        // a -> b -> c -> d: include a with deps, exclude c with preserve connectivity
+        let g = make_graph(
+            &[("a", "a"), ("b", "b"), ("c", "c"), ("d", "d")],
+            &[("a", "b"), ("b", "c"), ("c", "d")],
+            vec![],
+        );
+        let args = SelectArgs::default()
+            .include("a")
+            .deps()
+            .exclude("c")
+            .preserve_connectivity();
+        let result = select(&g, &args).unwrap();
+        assert_eq!(node_ids(&result), vec!["a", "b", "d"]);
+        assert_eq!(edge_pairs(&result), vec![("a", "b"), ("b", "d")]);
     }
 }
